backend-api-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
groq==0.4.1
httpx==0.25.2
pydantic==2.5.2
python-multipart==0.0.6
aiologger==0.7.0